roles creados definiendo los permisos, se anexo el boton habilitar y el campo sucursal en usuarios

+++ resources/views/tramiteshabilitados/index.blade.php
@@ -4,7 +4,6 @@
 <!-- page content -->
 
 <div class="container">
-
     <div class="row">
         <div class="col-sm-8 col-xs-8">
             {!! Form::open(['method'=>'GET','url'=>'tramitesHabilitados','class'=>'navbar-form navbar-left','role'=>'search'])  !!}
@@ -17,9 +16,11 @@
             {!! Form::close() !!}
         </div>
 
+        @can('add_tramitesHabilitados')
         <div class="col-sm-4 col-xs-4 text-right">
             <a href="{{route('tramitesHabilitados.create')}}" class="btn btn-primary">Nuevo <i class="glyphicon glyphicon-plus-sign"></i> </a>
         </div>
+        @endcan
     </div>
 
     <div class="table-responsive">
@@ -34,6 +35,7 @@
                     <th class="column-title">Pais</th>
                     <th class="column-title">Fecha</th>
                     <th class="column-title">Usuario</th>
+                    <th class="column-title">Habilitado</th>
                     <th class="column-title"></th>
                 </tr>
             </thead>
@@ -46,12 +48,24 @@
                     <td>{{ $row->nro_doc }}</td>
                     <td>{{ $row->pais }}</td>
                     <td>{{ $row->fecha }}</td>
-                    <td>{{ $row->user_id }}</td>                        
+                    <td>{{ $row->user_id }}</td>
                     <td>
-                        <a href="{{ route('tramitesHabilitados.edit', $row->id) }}" class="btn btn-success pull-right btn-xs" title="Editar"> Editar <i class="fa fa-edit"></i></a>                        
-                        {!! Form::open(array('route' => array('tramitesHabilitados.destroy', $row->id), 'method' => 'delete')) !!}
-                            <button class='btn btn-danger pull-right btn-xs' type="submit"> Borrar <i class="fa fa-trash"></i> </button>
-                        {!! Form::close() !!}
+                        @if($row->habilitado)
+                            <input id="habilitado{{ $row->id }}" type="checkbox" checked disabled data-toggle="toggle" data-on="Si" data-off="No" data-onstyle="success" data-offstyle="danger" data-size="mini" data-width="70" >
+                        @else
+                            <input id="habilitado{{ $row->id }}" type="checkbox" onchange="habilitarTurno({{ $row->id }})" data-toggle="toggle"  data-on="Si" data-off="No" data-onstyle="success" data-offstyle="danger" data-size="mini" data-width="70" >
+                        @endif
+                    </td>                                                
+                    <td>
+                        @can('edit_tramitesHabilitados')
+                            <a href="{{ route('tramitesHabilitados.edit', $row->id) }}" class="btn btn-success pull-right btn-xs" title="Editar"> Editar <i class="fa fa-edit"></i></a>
+                        @endcan
+
+                        @can('delete_tramitesHabilitados')
+                            {!! Form::open(array('route' => array('tramitesHabilitados.destroy', $row->id), 'method' => 'delete')) !!}
+                                <button class='btn btn-danger pull-right btn-xs' type="submit"> Borrar <i class="fa fa-trash"></i> </button>
+                            {!! Form::close() !!}
+                        @endcan
                     </td>
                 </tr>
             @endforeach
@@ -69,8 +83,28 @@
 <!-- /page content -->
 @endsection
 
+
 @push('scripts')
-  <script src="{{ asset('vendors/jquery/dist/jquery.min.js')}}"></script>
-  <!-- Bootstrap -->
-  <script src="{{ asset('vendors/bootstrap/dist/js/bootstrap.min.js')}}"></script>
+    <!-- Bootstrap-toggle -->
+    <script src="{{ asset('vendors/bootstrap-toggle/js/bootstrap-toggle.min.js')}}"></script>
+
+    <script>
+        function habilitarTurno(id){
+            var valor = $("#habilitado"+id).prop('checked');
+            $.ajax({
+                headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') },
+                url: '/tramitesHabilitadosHabilitar',
+                data: {id: id, valor:valor },
+                type: "GET", 
+                success: function(ret){
+                    //window.location.href = "{{ route('tramitesHabilitados.index') }}"; //using a named route
+                }
+            });
+        }
+    </script>
 @endpush
+
+@section('css')
+    <!-- bootstrap-toggle -->
+    <link href="{{ asset('vendors/bootstrap-toggle/css/bootstrap-toggle.min.css')}}" rel="stylesheet">
+@endsection
